pyproject.toml
[project]
name = "unity-mcp-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.20",
    "httpx>=0.27",
    "aiosqlite>=0.20",
    "gitpython>=3.1",
]

[project.optional-dependencies]
chroma = ["chromadb>=0.5"]